Cleanup: fix signed/unsigned comparison in ParticleTrail::ChangeMax
Cast gl_options.max_array_vertices/indices (int) to unsigned to match the
unsigned max*per_particle operands. Value is checked >0 first so the cast
is safe. Eliminates 2 -Wsign-compare warnings.

+++ src/gfx/particle.cpp
@@ -36,14 +36,14 @@ void ParticleTrail::ChangeMax( unsigned int max )
         vertices_per_particle = indices_per_particle = 12;
     
     if (gl_options.max_array_vertices > 0 && 
-            (max * vertices_per_particle) > gl_options.max_array_vertices)
+            (max * vertices_per_particle) > (unsigned int)gl_options.max_array_vertices)
     {
         max = gl_options.max_array_vertices / vertices_per_particle;
         if (max > 0)
             max--; // for a margin
     }
     if (gl_options.max_array_indices > 0 && 
-            (max * indices_per_particle) > gl_options.max_array_indices)
+            (max * indices_per_particle) > (unsigned int)gl_options.max_array_indices)
     {
         max = gl_options.max_array_indices / indices_per_particle;
         if (max > 0)
